refactor(web): adapt api specification; TODO: watched endpoints

+++ web/src/features/watchlist/api/remove-movie-from-watchlist.ts
@@ -1,20 +1,26 @@
 import { apiClient } from '@/lib/api-client';
 import { MutationConfig } from '@/lib/react-query';
-import { getMoviesWatchlistQueryOptions } from './get-movies-watchlist';
+import { isMovieOnWatchlistQueryOptions } from './is-movie-on-watchlist';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { getWatchlistQueryOptions } from './get-watchlist';
 
-export const removeMovieFromWatchlist = ({ id }: { id: number }) => {
+export const removeMovieFromWatchlist = (id: string) => {
+  if (!id) {
+    return Promise.reject(new Error('Movie id is required'));
+  }
   const url = `/watchlist/movies/${id}`;
   return apiClient.delete(url, true);
 };
 
 type UseRemoveMovieFromWatchlistOptions = {
+  id: string;
   mutationConfig?: MutationConfig<typeof removeMovieFromWatchlist>;
 };
 
 export const useRemoveMovieFromWatchlist = ({
+  id,
   mutationConfig,
-}: UseRemoveMovieFromWatchlistOptions = {}) => {
+}: UseRemoveMovieFromWatchlistOptions) => {
   const queryClient = useQueryClient();
 
   const { onSuccess, ...restConfig } = mutationConfig || {};
@@ -24,7 +30,10 @@ export const useRemoveMovieFromWatchlist = ({
     mutationFn: removeMovieFromWatchlist,
     onSuccess: (...args) => {
       queryClient.invalidateQueries({
-        queryKey: getMoviesWatchlistQueryOptions().queryKey,
+        queryKey: isMovieOnWatchlistQueryOptions({ id }).queryKey,
+      });
+      queryClient.invalidateQueries({
+        queryKey: getWatchlistQueryOptions().queryKey,
       });
       onSuccess?.(...args);
     },
